refactor: Modularize `app.js` codes
