+++ redis/Redis.js
@@ -1,0 +1,17 @@
+const redis = require('redis');
+
+
+// Redis 연결
+const redisClient = redis.createClient({
+  url: `redis://${process.env.REDIS_HOST}:${process.env.REDIS_PORT}`,
+  password: process.env.REDIS_PASSWORD,
+  legacyMode: false, // Promise
+});
+redisClient.on('connect', () => {
+  console.info('Redis connected!');
+});
+redisClient.on('error', (err) => {
+  console.error('Redis Client Error', err);
+});
+
+module.exports = redisClient;
